move types to eogrow.types
