+++ eogrow/core/storage.py
@@ -8,7 +8,7 @@ from pydantic import BaseSettings, Field
 from eolearn.core.utils.fs import get_aws_credentials, get_filesystem, is_s3_path
 from sentinelhub import SHConfig
 
-from ..utils.types import AwsAclType
+from ..types import AwsAclType
 from .base import EOGrowObject
 from .schemas import ManagerSchema
