+++ eogrow/pipelines/byoc.py
@@ -17,7 +17,7 @@ from sentinelhub.geometry import Geometry
 from eogrow.core.config import RawConfig
 
 from ..core.pipeline import Pipeline
-from ..utils.types import JsonDict
+from ..types import JsonDict
 from ..utils.validators import (
     ensure_defined_together,
     ensure_exactly_one_defined,
