+++ eogrow/pipelines/export_maps.py
@@ -25,8 +25,8 @@ from sentinelhub import CRS, MimeType
 from eogrow.core.config import RawConfig
 
 from ..core.pipeline import Pipeline
+from ..types import Feature
 from ..utils.map import CogifyResamplingOptions, WarpResamplingOptions, cogify_inplace, extract_bands, merge_tiffs
-from ..utils.types import Feature
 
 LOGGER = logging.getLogger(__name__)
