+++ eogrow/pipelines/features.py
@@ -27,8 +27,8 @@ from ..tasks.features import (
     ValidDataFractionPredicate,
     join_valid_and_cloud_masks,
 )
+from ..types import Feature, FeatureSpec, TimePeriod
 from ..utils.filter import get_patches_with_missing_features
-from ..utils.types import Feature, FeatureSpec, TimePeriod
 from ..utils.validators import field_validator, optional_field_validator, parse_dtype, parse_time_period
 
 LOGGER = logging.getLogger(__name__)
