+++ eogrow/pipelines/import_tiff.py
@@ -12,8 +12,8 @@ from eolearn.io import ImportFromTiffTask
 
 from ..core.pipeline import Pipeline
 from ..core.schemas import BaseSchema
+from ..types import Feature
 from ..utils.filter import get_patches_with_missing_features
-from ..utils.types import Feature
 from ..utils.validators import optional_field_validator, parse_dtype
