+++ eogrow/tasks/testing.py
@@ -7,7 +7,7 @@ import numpy as np
 from eolearn.core import EOPatch, EOTask, FeatureTypeSet
 from eolearn.core.utils.common import is_discrete_type
 
-from ..utils.types import Feature, TimePeriod
+from ..types import Feature, TimePeriod
 
 
 class DummyRasterFeatureTask(EOTask):
