+++ eogrow/utils/filter.py
@@ -10,9 +10,9 @@ from tqdm.auto import tqdm
 
 from eolearn.core import FeatureType
 from eolearn.core.eodata_io import walk_filesystem
-from eolearn.core.utils.types import EllipsisType
+from eolearn.core.types import EllipsisType
 
-from ..utils.types import FeatureSpec
+from ..types import FeatureSpec
 
 
 def check_if_features_exist(
