+++ eogrow/utils/testing.py
@@ -19,8 +19,8 @@ from sentinelhub import BBox
 
 from ..core.config import collect_configs_from_path, interpret_config_from_dict
 from ..core.pipeline import Pipeline
+from ..types import JsonDict
 from ..utils.meta import load_pipeline_class
-from ..utils.types import JsonDict
 
 
 class ContentTester:
