+++ eogrow/utils/types.py
@@ -1,37 +1,10 @@
-""" Includes custom types used in schemas
-"""
-import datetime
-from enum import Enum
-from typing import Any, Dict, Literal, Tuple, Union
+"""Deprecated module for types, moved to `eogrow.types`."""
+from warnings import warn
 
-from eolearn.core import FeatureType
+from ..types import *  # noqa # pylint: disable=wildcard-import,unused-wildcard-import
 
-Path = str
-S3Path = str
-ImportPath = str
-TimePeriod = Tuple[datetime.date, datetime.date]
-
-Feature = Tuple[FeatureType, str]
-FeatureSpec = Union[Tuple[FeatureType, str], FeatureType]
-
-BoolOrAuto = Union[Literal["auto"], bool]
-
-JsonDict = Dict[str, Any]
-RawSchemaDict = Dict[str, Any]
-
-AwsAclType = Literal[
-    "private",
-    "public-read",
-    "public-read-write",
-    "aws-exec-read",
-    "authenticated-read",
-    "bucket-owner-read",
-    "bucket-owner-full-control",
-    "log-delivery-write",
-]
-
-
-class ProcessingType(Enum):
-    RAY = "ray"
-    SINGLE = "single"
-    MULTI = "multi"
+warn(
+    "The module `eogrow.types` is deprecated, use `eogrow.types` instead.",
+    category=DeprecationWarning,
+    stacklevel=2,
+)
