+++ tests/test_utils/test_validators.py
@@ -10,7 +10,7 @@ from sentinelhub import DataCollection
 from sentinelhub.data_collections_bands import Band, MetaBands, Unit
 
 from eogrow.core.schemas import BaseSchema, ManagerSchema
-from eogrow.utils.types import RawSchemaDict
+from eogrow.types import RawSchemaDict
 from eogrow.utils.validators import (
     ensure_defined_together,
     ensure_exactly_one_defined,
